Reduce event mining time to 25% of original time!
Remove the bottleneck Files.list() and stores the paths in a list before
executing the event miner in parallel.

This would exhaust the CPU and throttle the performance because of too many context switches. Instead of using all possible threads we utilize only 75% of it to keep the CPU utilization at a total of 80-90% while mining. Previously this was at 10-25%.