Further compress the weights and golds
We will compress their events individually to only capture the change in gold or weight instead of both. This reduces them by an additional 40%(weights) and 20%(gold). So gold can be reduced by a total of 5%! of the parsed data.

Fix bug with exported events always on one timestamp per event log